refactor: AppBottomBar 분리
